Implemented booking service

+++ java/src/main/java/centraluniversity/app/booking/repositories/RoomRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import centraluniversity.app.booking.models.rooms.Room;
 
 @Repository
-public interface RoomRepository extends JpaRepository<Room, Long> {
+public interface RoomRepository extends JpaRepository<Room, Integer> {
     Optional<Room> findByName(String name);
     List<Room> findByNameContaining(String namePart);
 }
